requirements/install.txt
django
django-rest-swagger
yaml
zope.dottedname
